refactor: Legacy Mapper 제거 후 DispatcherServlet 패키지 이동
