+++ app/src/main/java/com/techcourse/Application.java
@@ -1,10 +1,9 @@
 package com.techcourse;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.io.IOException;
 import java.util.stream.Stream;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class Application {
 
@@ -14,9 +13,8 @@ public class Application {
 
     public static void main(final String[] args) throws Exception {
         final int port = defaultPortIfNull(args);
-        final var tomcat = new TomcatStarter(port);
+        final var tomcat = new TomcatStarter(port, Application.class);
         log.info("configuring app with basedir: {}", TomcatStarter.WEBAPP_DIR_LOCATION);
-
         tomcat.start();
         stop(tomcat);
     }
