--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/Controller.java
@@ -1,8 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.asis;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public interface Controller {
-    String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception;
-}
