+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/DispatcherServlet.java
@@ -1,8 +1,5 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerManager;
-import com.interface21.webmvc.servlet.mvc.tobe.ServletRequestHandler;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
@@ -15,17 +12,15 @@ public class DispatcherServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
-    private HandlerManager handlerManager;
+    private final AnnotationHandlerMapping annotationHandlerMapping;
 
-    public DispatcherServlet() {
+    public DispatcherServlet(Object applicationPath) {
+        this.annotationHandlerMapping = new AnnotationHandlerMapping(applicationPath);
     }
 
     @Override
     public void init() {
-        ServletRequestHandler annotationHandlerMapping = new AnnotationHandlerMapping(Application.class);
-        ServletRequestHandler manualHandlerMapping = new ManualHandlerMapping();
-        handlerManager = new HandlerManager(annotationHandlerMapping, manualHandlerMapping);
-        handlerManager.initialize();
+        annotationHandlerMapping.initialize();
     }
 
     @Override
@@ -34,7 +29,7 @@ public class DispatcherServlet extends HttpServlet {
         final String requestURI = request.getRequestURI();
         log.debug("Method : {}, Request URI : {}", request.getMethod(), requestURI);
         try {
-            handlerManager.handle(request, response);
+            annotationHandlerMapping.handle(request, response);
         } catch (Throwable e) {
             log.error("Exception : {}", e.getMessage(), e);
             throw new ServletException(e.getMessage());
