--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/ServletRequestHandler.java
@@ -1,13 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public interface ServletRequestHandler {
-
-    boolean canHandle(HttpServletRequest request);
-
-    void handle(HttpServletRequest request, HttpServletResponse response) throws Exception;
-
-    void initialize();
-}
